favoristes page and proper setup with modules
