+++ js13kpwa/index.html
@@ -25,6 +25,7 @@
 	<h1>MFW XLVII-Planner</h1>
 	<p class="description">Inoffizieler  <a href="https://musikfestwochen.ch/">Musikfestwochen</a> Planner</p>
 	<button id="notifications">Request dummy notifications</button>
+	<a href="./favorites.html"><button id="notifications">Merkliste</button></a>
 	<section id="content-main"></section>
 </main>
 <!-- The Modal -->
